- now copying rootmap and root pcm files to lib directory - fixed data readers to open TTrees with name pndsim instead of cbmsim - fixed bug in createLmdFitData reader when not using the filepath flag -f - introduced misalignment by transforming reco data points.   Also modified simulation scripts to correctly handle misalignment

+++ scripts/runLumiFullSimPixel.sh
@@ -26,8 +26,12 @@ start_evt=$((${num_evts}*${filename_index})) #number of events * filename index 
 
 #switch on "missing plane" search algorithm
 misspl=true
+
 #use cuts during trk seacrh with "CA". Should be 'false' if sensors missaligned!
 trkcut=true
+if [ "${alignment_matrices_path}" == "" ]; then
+  trkcut=false
+fi
 #merge hits on sensors from different sides. true=yes
 mergedHits=true
 #Skip kinematic filter (before back-propagation)
@@ -91,7 +95,7 @@ if [ 0 -eq "$?" ] || [ 2 -eq "${force_level}" ]; then
     #simulation with Box generator cheating with neutrons, since "no tracks" gave problems
     root -l -b -q 'runLumiPixel0SimBox.C('${num_evts}','${start_evt}',"'${workpathname}'",'$verbositylvl',2112,'${mom}','${numTrks}','${random_seed}')' > /dev/null 2>&1
   else
-    root -l -b -q 'runLumiPixel0SimDPM.C('${num_evts}','${start_evt}','${mom}',"'${gen_filepath}'", "'${workpathname}'",'$beamX0', '$beamY0', '${targetZ0}', '${beam_widthX}', '${beam_widthY}', '${target_widthZ}', '${beam_gradX}', '${beam_gradY}', '${beam_grad_sigmaX}', '${beam_grad_sigmaY}', "'${lmd_geometry_filename}'", "'${misalignment_matrices_path}'", '$verbositylvl')'
+    root -l -b -q 'runLumiPixel0SimDPM.C('${num_evts}','${start_evt}','${mom}',"'${gen_filepath}'", "'${workpathname}'",'$beamX0', '$beamY0', '${targetZ0}', '${beam_widthX}', '${beam_widthY}', '${target_widthZ}', '${beam_gradX}', '${beam_gradY}', '${beam_grad_sigmaX}', '${beam_grad_sigmaY}', "'${lmd_geometry_filename}'", "'${misalignment_matrices_path}'", '${use_point_transform_misalignment}', '$verbositylvl')'
   fi
   if [ "${debug}" -eq 0 ]; then 
     cp $workpathname/Lumi_MC_${start_evt}.root ${path_mc_data}/Lumi_MC_${start_evt}.root
@@ -110,20 +114,23 @@ if [ 0 -eq "$?" ] || [ 2 -eq "${force_level}" ]; then
   if [ ${simulate_noise} ]; then
     root -l -b -q 'runLumiPixel1bDigiNoise.C('${num_evts}','${start_evt}',"'${workpathname}'",'$verbositylvl', '${random_seed}')'
   else 
-    root -l -b -q 'runLumiPixel1Digi.C('${num_evts}','${start_evt}',"'${workpathname}'", "'${misalignment_matrices_path}'", '$verbositylvl')'
+    root -l -b -q 'runLumiPixel1Digi.C('${num_evts}','${start_evt}',"'${workpathname}'", "'${misalignment_matrices_path}'", '${use_point_transform_misalignment}', '$verbositylvl')'
   fi
 fi
 
 check_stage_success "$workpathname/Lumi_reco_${start_evt}.root"
 if [ 0 -eq "$?" ] || [ 2 -eq "${force_level}" ]; then
-  root -l -b -q 'runLumiPixel2Reco.C('${num_evts}','${start_evt}',"'${workpathname}'", "'${alignment_matrices_path}'", '$verbositylvl')'
+  root -l -b -q 'runLumiPixel2Reco.C('${num_evts}','${start_evt}',"'${workpathname}'", "'${alignment_matrices_path}'", "'${misalignment_matrices_path}'", '${use_point_transform_misalignment}', '$verbositylvl')'
 fi
 
 #find pairs
-root -l -b -q 'runLumiPixel2ePairFinder.C('${num_evts}','${start_evt}',"'${workpathname}'",'$verbositylvl')'
-#copy pairs
-if [ "${debug}" -eq 0 ]; then
-  cp $workpathname/Lumi_Pairs_${start_evt}.root ${pathname}/Pairs/Lumi_Pairs_${start_evt}.root
+check_stage_success "$workpathname/Lumi_Pairs_${start_evt}.root"
+if [ 0 -eq "$?" ] || [ 2 -eq "${force_level}" ]; then
+  root -l -b -q 'runLumiPixel2ePairFinder.C('${num_evts}','${start_evt}',"'${workpathname}'",'$verbositylvl')'
+  #copy pairs
+  if [ "${debug}" -eq 0 ]; then
+    cp $workpathname/Lumi_Pairs_${start_evt}.root ${pathname}/Pairs/Lumi_Pairs_${start_evt}.root
+  fi
 fi
 
 #merge hits
@@ -182,7 +189,7 @@ fi
 
 
 # filter back-propagated tracks (momentum cut)
-if [ $CleanSig == "true" ] || [ 2 -eq "${force_level}" ]; then 
+if [ $CleanSig == "true" ]; then 
   root -l -b -q 'runLumiPixel5bCleanSig.C('${num_evts}', '${start_evt}', "'${workpathname}'", '$verbositylvl', '${mom}', '${rec_ipx}', '${rec_ipy}')'
 fi
 
@@ -193,7 +200,7 @@ fi
 # this is required for the acceptance calculation
 check_stage_success "$workpathname/Lumi_TrksQA_${start_evt}.root"
 if [ 0 -eq "$?" ] || [ 2 -eq "${force_level}" ]; then
-  root -l -b -q 'runLumiPixel7TrksQA.C('${num_evts}','${start_evt}',"'${workpathname}'",'0','${mom}', '$WrAllMC', '${CleanSig}')'
+  root -l -b -q 'runLumiPixel7TrksQA.C('${num_evts}','${start_evt}',"'${workpathname}'",'$verbositylvl','${mom}', '$WrAllMC', '${CleanSig}')'
   if [ "${debug}" -eq 0 ]; then
     cp $workpathname/Lumi_TrksQA_${start_evt}.root $pathname/Lumi_TrksQA_${start_evt}.root
   fi
